requirements.txt
tensorflow>=2.10.0
keras>=2.10.0
pillow>=9.0.0
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0

scikit-learn>=1.0.0
tqdm>=4.60.0
